+++ foundation-models/openai/src/main/java/com/sap/ai/sdk/foundationmodels/openai/model/OpenAiChatCompletionOutput.java
@@ -1,7 +1,6 @@
 package com.sap.ai.sdk.foundationmodels.openai.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.google.common.annotations.Beta;
 import com.sap.ai.sdk.foundationmodels.openai.OpenAiClientException;
 import java.util.ArrayList;
 import java.util.List;
@@ -16,7 +15,7 @@ import lombok.experimental.Accessors;
 @Accessors(chain = true)
 @EqualsAndHashCode(callSuper = true)
 @ToString(callSuper = true)
-@Beta
+@Deprecated
 public class OpenAiChatCompletionOutput extends OpenAiCompletionOutput
     implements DeltaAggregatable<OpenAiChatCompletionDelta> {
   /** List of result candidates. */